Move the pre-existing functional interfaces back to the Functions class.
Also remove wildcard imports.
